// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  CUSTOMER: 'CUSTOMER',
  MANAGER: 'MANAGER',
  ADMIN: 'ADMIN',
  VENDOR: 'VENDOR',
  DELIVERY_MAN: 'DELIVERY_MAN',
  WAREHOUSE_MANAGER: 'WAREHOUSE_MANAGER',
  MANUAL_ORDER_ENTRY: 'MANUAL_ORDER_ENTRY'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const OrderStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  RETURNED: 'RETURNED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentGateway = {
  COD: 'COD',
  BKASH: 'BKASH',
  NAGAD: 'NAGAD',
  SSLCOMMERZ: 'SSLCOMMERZ',
  AAMARPAY: 'AAMARPAY',
  STRIPE: 'STRIPE'
} as const

export type PaymentGateway = (typeof PaymentGateway)[keyof typeof PaymentGateway]


export const DiscountType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED_AMOUNT: 'FIXED_AMOUNT'
} as const

export type DiscountType = (typeof DiscountType)[keyof typeof DiscountType]


export const InventoryAction = {
  STOCK_IN: 'STOCK_IN',
  STOCK_OUT: 'STOCK_OUT',
  ORDER_PLACED: 'ORDER_PLACED',
  ORDER_CANCELLED: 'ORDER_CANCELLED',
  RETURN_RECEIVED: 'RETURN_RECEIVED',
  DAMAGED: 'DAMAGED'
} as const

export type InventoryAction = (typeof InventoryAction)[keyof typeof InventoryAction]


export const ProductUnit = {
  KG: 'KG',
  GM: 'GM',
  MG: 'MG',
  LITER: 'LITER',
  ML: 'ML',
  PIECE: 'PIECE',
  PACKET: 'PACKET',
  BOX: 'BOX',
  DOZEN: 'DOZEN',
  PAIR: 'PAIR',
  BUNDLE: 'BUNDLE',
  BAG: 'BAG',
  BOTTLE: 'BOTTLE',
  CAN: 'CAN'
} as const

export type ProductUnit = (typeof ProductUnit)[keyof typeof ProductUnit]


export const ExpenseType = {
  RENT: 'RENT',
  SALARY: 'SALARY',
  UTILITY: 'UTILITY',
  PACKAGING: 'PACKAGING',
  MARKETING: 'MARKETING',
  DELIVERY_CHARGE: 'DELIVERY_CHARGE',
  DAMAGED_PRODUCT: 'DAMAGED_PRODUCT',
  OTHERS: 'OTHERS'
} as const

export type ExpenseType = (typeof ExpenseType)[keyof typeof ExpenseType]